requirements.txt
certifi==2017.11.5
future==0.16.0
python-telegram-bot==8.1.1
redis==2.10.5
